Map ITRS frames to terrestrial WCS coordinates
This will make it possible to use WCSAxes to make figures that
combine both celestial and terrestrial features. An example is
plotting the coordinates of an astronomical transient over an all-
sky satellite image to illustrate the position relative to the
Earth at the time of the event.

The ITRS frame is identified with WCSs that use the `TLON-` and
`TLAT-` coordinate types. There are several examples of WCSs where
this syntax is used to describe terrestrial coordinate systems:
Section 7.4.1 of the WCS in FITS "Paper II", and the WCSTools docs:
http://tdc-www.harvard.edu/software/wcstools/wcstools.multiwcs.html

This is the technique that I am using to make images such as this one:
http://ligo.org/detections/GW170817/images-GW170817/skymap+earth.jpg